finishing test email confirmation
