+++ src/utils/createConfirmedEmailLink.ts
@@ -1,29 +1,8 @@
-import * as Redis from "ioredis";
-import fetch from "node-fetch";
-import { createTypeormConnection } from "./createTypeormConnection";
-import { User } from "../entity/User";
-import { createConfirmedEmailLink } from "./createConfirmeEmailLink";
+import {v4} from "uuid";
+import { Redis } from "ioredis";
 
-let userId = "";
-
-beforeAll(async () => {
-  await createTypeormConnection();
-  const user = await User.create({
-    email: "zei1@han.com",
-    password: "asdasdasd"
-  }).save();
-
-  userId = user.id;
-});
-
-test("Make sure createConfirmedEmailLink works", async () => {
-  const url = await createConfirmedEmailLink(
-    process.env.TEST_HOST as string,
-    userId,
-    new Redis()
-  );
-
-  const response = await fetch(url);
-  const text = await response.text();
-  console.log(text);
-});
+export const createConfirmedEmailLink = async (url: string, userId: string, redis: Redis) => {
+    const id = v4();
+    await redis.set(id, userId, "ex", 60*60*24);
+    return `${url}/confirm/${id}`;
+}
